requirements.txt
chardet==5.0.0
pdfminer.six==20200517
pdfplumber==0.5.23
Pillow==9.2.0
pycryptodome==3.15.0
python-dateutil==2.8.2
six==1.16.0
sortedcontainers==2.4.0
Wand==0.6.7
